1.Fix bugs in compressing numbers of small stream fragments 2.Minor refactor, fix a possible race-condition bug on recyling the socket 3.Protocol glitch fixed:   now take the salt generated by the initiator as the leftmost 32 bits of the 96-bit nonce   make it NAT-friendly when underlying packet transfer protocol is UDP/IPv4 4.API glitch fixed: GET/SET EXTent POINTER 5.Minor bug fixed in acquire cryptography context under Windows   exploit CRYPT_VERIFYCONTEXT flag to avoid accessing private key set   and eliminate private-key related exception 6.Fix bug in handling connection multiplication   Protocol glitch fixed: clarify constitution of the nonce   exploited to in get/verify ICC of the MULTIPLY packet 7.Minor refactor, put packet structure definition in FSP.h 8.Prepare to separate implementation of FSP over IPv6 and FSP over UDP/IPv4

+++ src/FileSyncServer/MemoryFile.cpp
@@ -19,7 +19,7 @@ void SendMemoryPattern()
 		* (uint32_t *) & bytesToSend[i * sizeof(uint32_t)] = htobe32(i);
 	}
 
-	WaitConnection(defaultWelcome, strlen(defaultWelcome) + 1, onAccepted);
+	WaitConnection(defaultWelcome, (uint16_t)strlen(defaultWelcome) + 1, onAccepted);
 }
